pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Aaron Nicholas Gumapac <aarn.gmpc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.23"
python-jose = "^3.3.0"
motor = "^3.3.2"
pydantic-settings = "^2.1.0"
pymongo = "^4.6.1"
cryptography = "^41.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
